+++ integrations/cron/v3/template.go
@@ -8,6 +8,7 @@ type Template interface {
 func BuiltinTemplates() []Template {
     return []Template{
         defaultCrontabTemplate,
+        defaultK8sTemplate,
     }
 }