+++ v3/mailer/contract/mailer.go
@@ -13,6 +13,8 @@ type Attachment struct {
     Filename    string
     ContentType string
     Content     []byte
+    /* @info when ContentId is non-empty the attachment is embedded inline (Content-Disposition: inline) inside a multipart/related part so an HTML body can reference it as <img src="cid:..."> */
+    ContentId string
 }
 
 type Message struct {